warehouse-api/requirements.txt
fastapi>=0.100.0,<0.101.0
uvicorn[standard]>=0.22.0,<0.23.0
gunicorn
fastapi-users>=12.1.2,<12.2.0
httpx-oauth>=0.10.2,<0.11.0
fastapi-users-db-sqlalchemy>=6.0.1,<6.1.0
pydantic>=2,<3
pydantic-settings>=2,<3
yarl>=1.9.2,<1.12.1
ujson>=5.8.0,<5.9.0
SQLAlchemy[asyncio]>=2.0.18,<2.1.0
alembic>=1.11.1,<1.12.0
asyncpg[sa]>=0.28.0,<0.29.0
redis[hiredis]
aiofiles>=23.1.0,<24.0.0
httptools>=0.6.0,<0.7.0
strawberry-graphql[fastapi]>=0.194.4,<0.195.0
aio-pika>=9.1.4,<9.2.0
prometheus-client>=0.17.0,<0.18.0
prometheus-fastapi-instrumentator==6.0.0
opentelemetry-api>=1.18.0,<1.19.0
opentelemetry-sdk>=1.18.0,<1.19.0
opentelemetry-exporter-otlp>=1.18.0,<1.19.0
opentelemetry-instrumentation>=0.39b0,<0.40b0
opentelemetry-instrumentation-fastapi>=0.39b0,<0.40b0
opentelemetry-instrumentation-redis>=0.39b0,<0.40b0
opentelemetry-instrumentation-sqlalchemy>=0.39b0,<0.40b0
opentelemetry-instrumentation-aio-pika>=0.39b0,<0.40b0
loguru>=0.7.2,<0.8.0
taskiq
taskiq-fastapi
taskiq-redis
taskiq-aio-pika
pytest
flake8~=4.0.1
mypy>=1.1.1,<1.2.0
isort>=5.11.4,<5.12.0
pre-commit>=3.0.1,<3.1.0
wemake-python-styleguide>=0.17.0,<0.18.0
black>=22.12.0,<22.13.0
autoflake>=1.6.1,<1.7.0
pytest-cov>=4.0.0,<4.1.0
anyio>=3.6.2,<3.7.0

pytest-env>=0.8.1,<0.9.0
fakeredis>=2.5.0,<2.6.0
httpx>=0.23.3,<0.24.0
taskiq[reload]
elasticsearch-dsl[async]==8.15.3
aiohttp[speedups]==3.10.8
